Ported Task Board Functionality
I just ported kendell's work into the react js taskboard component.

there is a slight problem where one line of code is not recognizable for the js file. I added a little comment indicating it. it is related to the functionality that allows kendell's prototype to save memory between site visits, which is broken for the project version